Created Animal Prefabs and Spawning
Created Cow, Horse, and Moose animal prefabs. Created a movement script to move the animals forward. Created a spawn manager and spawn manager script to spawn animals randomly on the top, bottom, left, and right of the screen to charge across when the s key is pressed. Did not add functionality to delete animals when they go out of bounds yet